Integrate LocalStore into BTESecureStorage.
